removing duplication of code, moving common methods to helper file
